Expression data from healthy and malignant (chronic lymphocytic leukemia, CLL) human B-lymphocytes after B-cell receptor stimulation (GSE 39411 dataset)

B-cells were negatively selected from healthy donors and previously untreated CLL patients. BCR stimulated and unstimulated control B-cells were treated at four time points after stimulation for total RNA extraction and hybridization on Affymetrix microarrays.

Format

The format is: chr "CLL"

Source

https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE39411

Details

The dataset provided with package is the first five lines of the full dataset. The full dataset can be downloaded from the github repository of the package (https://raw.githubusercontent.com/fbertran/Patterns/master/add_data/CLL.RData).

Three different cell populations (6 healthy B-lymphocytes, 6 leukemic CLL B-lymphocyte of indolent form and 5 leukemic CLL B-lymphocyte of aggressive form) were stimulated in vitro with an anti-IgM antibody, activating the B-cell receptor (BCR). We analyzed the gene expression at 4 time points (60, 90, 210 and 390 minutes). Each gene expression measurement is performed both in stimulated cells and in control unstimulated cells. For one aggressive CLL case, we silenced expression of DUSP1 by transfecting DUSP1-specific RNAi and, as a control, transfected cells with a non-targeting RNAi. We then stimulated the BCR of these cells and analyzed the gene expression at the same time points in stimulated cells and in control unstimulated cells.

References

Vallat, L., Kemper, C. A., Jung, N., Maumy-Bertrand, M., Bertrand, F., Meyer, N., … Bahram, S. (2013). Reverse-engineering the genetic circuitry of a cancer cell with predicted intervention in chronic lymphocytic leukemia. Proceedings of the National Academy of Sciences of the United States of America, 110(2), 459–464.

Author

Bertrand Frederic, Myriam Maumy-Bertrand.

Examples

data(CLL)
str(CLL)
#> 'data.frame':	6 obs. of  154 variables:
#>  $ probeset                                                                             : chr  "1007_s_at" "1053_at" "117_at" "121_at" ...
#>  $ nom                                                                                  : chr  "DDR1 or MIR4640" "RFC2" "HSPA6" "PAX8" ...
#>  $ Healthy.B.cell..subject.N1_unstimulated.cells_at.60min                               : num  561 221 270 678 142 ...
#>  $ Healthy.B.cell..subject.N1_unstimulated.cells_at.90min                               : num  557 216 263 598 136 ...
#>  $ Healthy.B.cell..subject.N1_unstimulated.cells_at.210min                              : num  561 229 244 618 145 ...
#>  $ Healthy.B.cell..subject.N1_unstimulated.cells_at.390min                              : num  647 224 252 716 135 ...
#>  $ Healthy.B.cell..subject.N1_stimulated.cells_at.60min                                 : num  566 217 279 659 144 ...
#>  $ Healthy.B.cell..subject.N1_stimulated.cells_at.90min                                 : num  584 203 246 625 126 ...
#>  $ Healthy.B.cell..subject.N1_stimulated.cells_at.210min                                : num  522 210 241 590 140 ...
#>  $ Healthy.B.cell..subject.N1_stimulated.cells_at.390min                                : num  561 210 241 676 138 ...
#>  $ Healthy.B.cell..subject.N2_unstimulated.cells_at.60min                               : num  458 216 259 516 136 ...
#>  $ Healthy.B.cell..subject.N2_unstimulated.cells_at.90min                               : num  570 224 264 636 137 ...
#>  $ Healthy.B.cell..subject.N2_unstimulated.cells_at.210min                              : num  573 216 255 574 141 ...
#>  $ Healthy.B.cell..subject.N2_unstimulated.cells_at.390min                              : num  583 224 253 612 122 ...
#>  $ Healthy.B.cell..subject.N2_stimulated.cells_at.60min                                 : num  602 229 294 686 149 ...
#>  $ Healthy.B.cell..subject.N2_stimulated.cells_at.90min                                 : num  539 224 290 670 139 ...
#>  $ Healthy.B.cell..subject.N2_stimulated.cells_at.210min                                : num  531 220 261 585 140 ...
#>  $ Healthy.B.cell..subject.N2_stimulated.cells_at.390min                                : num  570 219 250 568 128 ...
#>  $ Healthy.B.cell..subject.N3_unstimulated.cells_at.60min                               : num  560 250 250 609 149 ...
#>  $ Healthy.B.cell..subject.N3_unstimulated.cells_at.90min                               : num  632 240 245 657 146 ...
#>  $ Healthy.B.cell..subject.N3_unstimulated.cells_at.210min                              : num  651 248 255 723 128 ...
#>  $ Healthy.B.cell..subject.N3_unstimulated.cells_at.390min                              : num  626 235 240 677 140 ...
#>  $ Healthy.B.cell..subject.N3_stimulated.cells_at.60min                                 : num  555 251 252 620 139 ...
#>  $ Healthy.B.cell..subject.N3_stimulated.cells_at.90min                                 : num  605 247 242 715 127 ...
#>  $ Healthy.B.cell..subject.N3_stimulated.cells_at.210min                                : num  449 227 236 479 124 ...
#>  $ Healthy.B.cell..subject.N3_stimulated.cells_at.390min                                : num  543 229 242 683 134 ...
#>  $ Healthy.B.cell..subject.N4_unstimulated.cells_at.60min                               : num  589 255 257 688 147 ...
#>  $ Healthy.B.cell..subject.N4_unstimulated.cells_at.90min                               : num  614 244 243 607 137 ...
#>  $ Healthy.B.cell..subject.N4_unstimulated.cells_at.210min                              : num  611 238 253 691 138 ...
#>  $ Healthy.B.cell..subject.N4_unstimulated.cells_at.390min                              : num  601 252 270 711 133 ...
#>  $ Healthy.B.cell..subject.N4_stimulated.cells_at.60min                                 : num  586 255 253 637 142 ...
#>  $ Healthy.B.cell..subject.N4_stimulated.cells_at.90min                                 : num  582 230 240 672 137 ...
#>  $ Healthy.B.cell..subject.N4_stimulated.cells_at.210min                                : num  537 236 255 675 140 ...
#>  $ Healthy.B.cell..subject.N4_stimulated.cells_at.390min                                : num  574 240 257 656 141 ...
#>  $ Healthy.B.cell..subject.N5_unstimulated.cells_at.60min                               : num  525 241 268 629 149 ...
#>  $ Healthy.B.cell..subject.N5_unstimulated.cells_at.90min                               : num  560 240 284 710 154 ...
#>  $ Healthy.B.cell..subject.N5_unstimulated.cells_at.210min                              : num  604 255 256 673 141 ...
#>  $ Healthy.B.cell..subject.N5_unstimulated.cells_at.390min                              : num  568 235 258 642 131 ...
#>  $ Healthy.B.cell..subject.N5_stimulated.cells_at.60min                                 : num  560 250 272 699 150 ...
#>  $ Healthy.B.cell..subject.N5_stimulated.cells_at.90min                                 : num  549 226 259 687 140 ...
#>  $ Healthy.B.cell..subject.N5_stimulated.cells_at.210min                                : num  563 231 280 683 146 ...
#>  $ Healthy.B.cell..subject.N5_stimulated.cells_at.390min                                : num  592 248 264 741 157 ...
#>  $ Healthy.B.cell..subject.N6_unstimulated.cells_at.60min                               : num  627 243 268 680 140 ...
#>  $ Healthy.B.cell..subject.N6_unstimulated.cells_at.90min                               : num  620 237 278 729 139 ...
#>  $ Healthy.B.cell..subject.N6_unstimulated.cells_at.210min                              : num  619 233 264 689 142 ...
#>  $ Healthy.B.cell..subject.N6_unstimulated.cells_at.390min                              : num  553 240 265 749 144 ...
#>  $ Healthy.B.cell..subject.N6_stimulated.cells_at.60min                                 : num  576 251 279 678 142 ...
#>  $ Healthy.B.cell..subject.N6_stimulated.cells_at.90min                                 : num  584 227 287 691 154 ...
#>  $ Healthy.B.cell..subject.N6_stimulated.cells_at.210min                                : num  593 225 268 719 148 ...
#>  $ Healthy.B.cell..subject.N6_stimulated.cells_at.390min                                : num  567 230 243 644 139 ...
#>  $ CLL.B.cell..patient.M1_unstimulated.cells_at.60min_.ind.                             : num  648 200 232 629 143 ...
#>  $ CLL.B.cell..patient.M1_unstimulated.cells_at.90min_.ind.                             : num  674 217 256 767 154 ...
#>  $ CLL.B.cell..patient.M1_unstimulated.cells_at.210min_.ind.                            : num  639 216 240 707 137 ...
#>  $ CLL.B.cell..patient.M1_unstimulated.cells_at.390min_.ind.                            : num  598 215 235 661 148 ...
#>  $ CLL.B.cell..patient.M1_stimulated.cells_at.60min_.ind.                               : num  609 201 234 724 138 ...
#>  $ CLL.B.cell..patient.M1_stimulated.cells_at.90min_.ind.                               : num  622 208 236 750 144 ...
#>  $ CLL.B.cell..patient.M1_stimulated.cells_at.210min_.ind.                              : num  598 193 231 721 145 ...
#>  $ CLL.B.cell..patient.M1_stimulated.cells_at.390min_.ind.                              : num  674 203 228 696 144 ...
#>  $ CLL.B.cell..patient.M4_unstimulated.cells_at.60min_.ind.                             : num  613 202 256 694 200 ...
#>  $ CLL.B.cell..patient.M4_unstimulated.cells_at.90min_.ind.                             : num  533 204 275 641 152 ...
#>  $ CLL.B.cell..patient.M4_unstimulated.cells_at.210min_.ind.                            : num  586 194 269 616 148 ...
#>  $ CLL.B.cell..patient.M4_unstimulated.cells_at.390min_.ind.                            : num  668 229 322 841 155 ...
#>  $ CLL.B.cell..patient.M4_stimulated.cells_at.60min_.ind.                               : num  560 201 225 676 153 ...
#>  $ CLL.B.cell..patient.M4_stimulated.cells_at.90min_.ind.                               : num  624 200 262 661 142 ...
#>  $ CLL.B.cell..patient.M4_stimulated.cells_at.210min_.ind.                              : num  542 229 318 703 153 ...
#>  $ CLL.B.cell..patient.M4_stimulated.cells_at.390min_.ind.                              : num  645 210 291 812 152 ...
#>  $ CLL.B.cell..patient.M5_unstimulated.cells_at.60min_.ind.                             : num  525 205 261 679 146 ...
#>  $ CLL.B.cell..patient.M5_unstimulated.cells_at.90min_.ind.                             : num  619 207 279 791 142 ...
#>  $ CLL.B.cell..patient.M5_unstimulated.cells_at.210min_.ind.                            : num  575 224 270 802 146 ...
#>  $ CLL.B.cell..patient.M5_unstimulated.cells_at.390min_.ind.                            : num  536 201 272 625 140 ...
#>  $ CLL.B.cell..patient.M5_stimulated.cells_at.60min_.ind.                               : num  605 206 292 829 148 ...
#>  $ CLL.B.cell..patient.M5_stimulated.cells_at.90min_.ind.                               : num  542 191 255 698 143 ...
#>  $ CLL.B.cell..patient.M5_stimulated.cells_at.210min_.ind.                              : num  510 200 267 692 145 ...
#>  $ CLL.B.cell..patient.M5_stimulated.cells_at.390min_.ind.                              : num  584 234 263 689 143 ...
#>  $ CLL.B.cell..patient.M6_unstimulated.cells_at.60min_.ind.                             : num  542 209 268 682 159 ...
#>  $ CLL.B.cell..patient.M6_unstimulated.cells_at.90min_.ind.                             : num  552 208 246 618 139 ...
#>  $ CLL.B.cell..patient.M6_unstimulated.cells_at.210min_.ind.                            : num  636 198 239 692 138 ...
#>  $ CLL.B.cell..patient.M6_unstimulated.cells_at.390min_.ind.                            : num  604 204 231 603 144 ...
#>  $ CLL.B.cell..patient.M6_stimulated.cells_at.60min_.ind.                               : num  567 199 277 686 139 ...
#>  $ CLL.B.cell..patient.M6_stimulated.cells_at.90min_.ind.                               : num  573 205 234 593 141 ...
#>  $ CLL.B.cell..patient.M6_stimulated.cells_at.210min_.ind.                              : num  652 203 234 577 147 ...
#>  $ CLL.B.cell..patient.M6_stimulated.cells_at.390min_.ind.                              : num  506 197 252 598 141 ...
#>  $ CLL.B.cell..patient.M7_unstimulated.cells_at.60min_.ind.                             : num  584 202 229 751 134 ...
#>  $ CLL.B.cell..patient.M7_unstimulated.cells_at.90min_.ind.                             : num  561 197 230 778 128 ...
#>  $ CLL.B.cell..patient.M7_unstimulated.cells_at.210min_.ind.                            : num  570 210 235 713 128 ...
#>  $ CLL.B.cell..patient.M7_unstimulated.cells_at.390min_.ind.                            : num  595 210 228 647 126 ...
#>  $ CLL.B.cell..patient.M7_stimulated.cells_at.60min_.ind.                               : num  528 187 218 591 134 ...
#>  $ CLL.B.cell..patient.M7_stimulated.cells_at.90min_.ind.                               : num  527 195 217 614 126 ...
#>  $ CLL.B.cell..patient.M7_stimulated.cells_at.210min_.ind.                              : num  653 216 220 737 130 ...
#>  $ CLL.B.cell..patient.M7_stimulated.cells_at.390min_.ind.                              : num  459 211 206 606 140 ...
#>  $ CLL.B.cell..patient.M8_unstimulated.cells_at.60min_.ind.                             : num  503 202 276 621 134 ...
#>  $ CLL.B.cell..patient.M8_unstimulated.cells_at.90min_.ind.                             : num  595 207 303 803 141 ...
#>  $ CLL.B.cell..patient.M8_unstimulated.cells_at.210min_.ind.                            : num  632 200 265 741 143 ...
#>  $ CLL.B.cell..patient.M8_unstimulated.cells_at.390min_.ind.                            : num  740 201 269 813 141 ...
#>  $ CLL.B.cell..patient.M8_stimulated.cells_at.60min_.ind.                               : num  493 192 301 677 132 ...
#>  $ CLL.B.cell..patient.M8_stimulated.cells_at.90min_.ind.                               : num  671 209 269 708 137 ...
#>  $ CLL.B.cell..patient.M8_stimulated.cells_at.210min_.ind.                              : num  721 213 259 848 143 ...
#>  $ CLL.B.cell..patient.M8_stimulated.cells_at.390min_.ind.                              : num  658 206 293 725 134 ...
#>  $ CLL.B.cell..patient.M2_unstimulated.cells_at.60min_.agg.                             : num  701 216 733 790 143 ...
#>   [list output truncated]

# \donttest{

CLLfile <- "https://github.com/fbertran/Patterns/raw/master/add_data/CLL.RData"
download.file(CLLfile,destfile = file.path(tempdir(),"downloadData.RData"),method = "libcurl")
load(file.path(tempdir(),"downloadData.RData"))
unlink(file.path(tempdir(),"downloadData.RData"))
str(CLL)
#> 'data.frame':	54675 obs. of  154 variables:
#>  $ probeset                                                                             : chr  "1007_s_at" "1053_at" "117_at" "121_at" ...
#>  $ nom                                                                                  : chr  "DDR1 or MIR4640" "RFC2" "HSPA6" "PAX8" ...
#>  $ Healthy.B.cell..subject.N1_unstimulated.cells_at.60min                               : num  561 221 270 678 142 ...
#>  $ Healthy.B.cell..subject.N1_unstimulated.cells_at.90min                               : num  557 216 263 598 136 ...
#>  $ Healthy.B.cell..subject.N1_unstimulated.cells_at.210min                              : num  561 229 244 618 145 ...
#>  $ Healthy.B.cell..subject.N1_unstimulated.cells_at.390min                              : num  647 224 252 716 135 ...
#>  $ Healthy.B.cell..subject.N1_stimulated.cells_at.60min                                 : num  566 217 279 659 144 ...
#>  $ Healthy.B.cell..subject.N1_stimulated.cells_at.90min                                 : num  584 203 246 625 126 ...
#>  $ Healthy.B.cell..subject.N1_stimulated.cells_at.210min                                : num  522 210 241 590 140 ...
#>  $ Healthy.B.cell..subject.N1_stimulated.cells_at.390min                                : num  561 210 241 676 138 ...
#>  $ Healthy.B.cell..subject.N2_unstimulated.cells_at.60min                               : num  458 216 259 516 136 ...
#>  $ Healthy.B.cell..subject.N2_unstimulated.cells_at.90min                               : num  570 224 264 636 137 ...
#>  $ Healthy.B.cell..subject.N2_unstimulated.cells_at.210min                              : num  573 216 255 574 141 ...
#>  $ Healthy.B.cell..subject.N2_unstimulated.cells_at.390min                              : num  583 224 253 612 122 ...
#>  $ Healthy.B.cell..subject.N2_stimulated.cells_at.60min                                 : num  602 229 294 686 149 ...
#>  $ Healthy.B.cell..subject.N2_stimulated.cells_at.90min                                 : num  539 224 290 670 139 ...
#>  $ Healthy.B.cell..subject.N2_stimulated.cells_at.210min                                : num  531 220 261 585 140 ...
#>  $ Healthy.B.cell..subject.N2_stimulated.cells_at.390min                                : num  570 219 250 568 128 ...
#>  $ Healthy.B.cell..subject.N3_unstimulated.cells_at.60min                               : num  560 250 250 609 149 ...
#>  $ Healthy.B.cell..subject.N3_unstimulated.cells_at.90min                               : num  632 240 245 657 146 ...
#>  $ Healthy.B.cell..subject.N3_unstimulated.cells_at.210min                              : num  651 248 255 723 128 ...
#>  $ Healthy.B.cell..subject.N3_unstimulated.cells_at.390min                              : num  626 235 240 677 140 ...
#>  $ Healthy.B.cell..subject.N3_stimulated.cells_at.60min                                 : num  555 251 252 620 139 ...
#>  $ Healthy.B.cell..subject.N3_stimulated.cells_at.90min                                 : num  605 247 242 715 127 ...
#>  $ Healthy.B.cell..subject.N3_stimulated.cells_at.210min                                : num  449 227 236 479 124 ...
#>  $ Healthy.B.cell..subject.N3_stimulated.cells_at.390min                                : num  543 229 242 683 134 ...
#>  $ Healthy.B.cell..subject.N4_unstimulated.cells_at.60min                               : num  589 255 257 688 147 ...
#>  $ Healthy.B.cell..subject.N4_unstimulated.cells_at.90min                               : num  614 244 243 607 137 ...
#>  $ Healthy.B.cell..subject.N4_unstimulated.cells_at.210min                              : num  611 238 253 691 138 ...
#>  $ Healthy.B.cell..subject.N4_unstimulated.cells_at.390min                              : num  601 252 270 711 133 ...
#>  $ Healthy.B.cell..subject.N4_stimulated.cells_at.60min                                 : num  586 255 253 637 142 ...
#>  $ Healthy.B.cell..subject.N4_stimulated.cells_at.90min                                 : num  582 230 240 672 137 ...
#>  $ Healthy.B.cell..subject.N4_stimulated.cells_at.210min                                : num  537 236 255 675 140 ...
#>  $ Healthy.B.cell..subject.N4_stimulated.cells_at.390min                                : num  574 240 257 656 141 ...
#>  $ Healthy.B.cell..subject.N5_unstimulated.cells_at.60min                               : num  525 241 268 629 149 ...
#>  $ Healthy.B.cell..subject.N5_unstimulated.cells_at.90min                               : num  560 240 284 710 154 ...
#>  $ Healthy.B.cell..subject.N5_unstimulated.cells_at.210min                              : num  604 255 256 673 141 ...
#>  $ Healthy.B.cell..subject.N5_unstimulated.cells_at.390min                              : num  568 235 258 642 131 ...
#>  $ Healthy.B.cell..subject.N5_stimulated.cells_at.60min                                 : num  560 250 272 699 150 ...
#>  $ Healthy.B.cell..subject.N5_stimulated.cells_at.90min                                 : num  549 226 259 687 140 ...
#>  $ Healthy.B.cell..subject.N5_stimulated.cells_at.210min                                : num  563 231 280 683 146 ...
#>  $ Healthy.B.cell..subject.N5_stimulated.cells_at.390min                                : num  592 248 264 741 157 ...
#>  $ Healthy.B.cell..subject.N6_unstimulated.cells_at.60min                               : num  627 243 268 680 140 ...
#>  $ Healthy.B.cell..subject.N6_unstimulated.cells_at.90min                               : num  620 237 278 729 139 ...
#>  $ Healthy.B.cell..subject.N6_unstimulated.cells_at.210min                              : num  619 233 264 689 142 ...
#>  $ Healthy.B.cell..subject.N6_unstimulated.cells_at.390min                              : num  553 240 265 749 144 ...
#>  $ Healthy.B.cell..subject.N6_stimulated.cells_at.60min                                 : num  576 251 279 678 142 ...
#>  $ Healthy.B.cell..subject.N6_stimulated.cells_at.90min                                 : num  584 227 287 691 154 ...
#>  $ Healthy.B.cell..subject.N6_stimulated.cells_at.210min                                : num  593 225 268 719 148 ...
#>  $ Healthy.B.cell..subject.N6_stimulated.cells_at.390min                                : num  567 230 243 644 139 ...
#>  $ CLL.B.cell..patient.M1_unstimulated.cells_at.60min_.ind.                             : num  648 200 232 629 143 ...
#>  $ CLL.B.cell..patient.M1_unstimulated.cells_at.90min_.ind.                             : num  674 217 256 767 154 ...
#>  $ CLL.B.cell..patient.M1_unstimulated.cells_at.210min_.ind.                            : num  639 216 240 707 137 ...
#>  $ CLL.B.cell..patient.M1_unstimulated.cells_at.390min_.ind.                            : num  598 215 235 661 148 ...
#>  $ CLL.B.cell..patient.M1_stimulated.cells_at.60min_.ind.                               : num  609 201 234 724 138 ...
#>  $ CLL.B.cell..patient.M1_stimulated.cells_at.90min_.ind.                               : num  622 208 236 750 144 ...
#>  $ CLL.B.cell..patient.M1_stimulated.cells_at.210min_.ind.                              : num  598 193 231 721 145 ...
#>  $ CLL.B.cell..patient.M1_stimulated.cells_at.390min_.ind.                              : num  674 203 228 696 144 ...
#>  $ CLL.B.cell..patient.M4_unstimulated.cells_at.60min_.ind.                             : num  613 202 256 694 200 ...
#>  $ CLL.B.cell..patient.M4_unstimulated.cells_at.90min_.ind.                             : num  533 204 275 641 152 ...
#>  $ CLL.B.cell..patient.M4_unstimulated.cells_at.210min_.ind.                            : num  586 194 269 616 148 ...
#>  $ CLL.B.cell..patient.M4_unstimulated.cells_at.390min_.ind.                            : num  668 229 322 841 155 ...
#>  $ CLL.B.cell..patient.M4_stimulated.cells_at.60min_.ind.                               : num  560 201 225 676 153 ...
#>  $ CLL.B.cell..patient.M4_stimulated.cells_at.90min_.ind.                               : num  624 200 262 661 142 ...
#>  $ CLL.B.cell..patient.M4_stimulated.cells_at.210min_.ind.                              : num  542 229 318 703 153 ...
#>  $ CLL.B.cell..patient.M4_stimulated.cells_at.390min_.ind.                              : num  645 210 291 812 152 ...
#>  $ CLL.B.cell..patient.M5_unstimulated.cells_at.60min_.ind.                             : num  525 205 261 679 146 ...
#>  $ CLL.B.cell..patient.M5_unstimulated.cells_at.90min_.ind.                             : num  619 207 279 791 142 ...
#>  $ CLL.B.cell..patient.M5_unstimulated.cells_at.210min_.ind.                            : num  575 224 270 802 146 ...
#>  $ CLL.B.cell..patient.M5_unstimulated.cells_at.390min_.ind.                            : num  536 201 272 625 140 ...
#>  $ CLL.B.cell..patient.M5_stimulated.cells_at.60min_.ind.                               : num  605 206 292 829 148 ...
#>  $ CLL.B.cell..patient.M5_stimulated.cells_at.90min_.ind.                               : num  542 191 255 698 143 ...
#>  $ CLL.B.cell..patient.M5_stimulated.cells_at.210min_.ind.                              : num  510 200 267 692 145 ...
#>  $ CLL.B.cell..patient.M5_stimulated.cells_at.390min_.ind.                              : num  584 234 263 689 143 ...
#>  $ CLL.B.cell..patient.M6_unstimulated.cells_at.60min_.ind.                             : num  542 209 268 682 159 ...
#>  $ CLL.B.cell..patient.M6_unstimulated.cells_at.90min_.ind.                             : num  552 208 246 618 139 ...
#>  $ CLL.B.cell..patient.M6_unstimulated.cells_at.210min_.ind.                            : num  636 198 239 692 138 ...
#>  $ CLL.B.cell..patient.M6_unstimulated.cells_at.390min_.ind.                            : num  604 204 231 603 144 ...
#>  $ CLL.B.cell..patient.M6_stimulated.cells_at.60min_.ind.                               : num  567 199 277 686 139 ...
#>  $ CLL.B.cell..patient.M6_stimulated.cells_at.90min_.ind.                               : num  573 205 234 593 141 ...
#>  $ CLL.B.cell..patient.M6_stimulated.cells_at.210min_.ind.                              : num  652 203 234 577 147 ...
#>  $ CLL.B.cell..patient.M6_stimulated.cells_at.390min_.ind.                              : num  506 197 252 598 141 ...
#>  $ CLL.B.cell..patient.M7_unstimulated.cells_at.60min_.ind.                             : num  584 202 229 751 134 ...
#>  $ CLL.B.cell..patient.M7_unstimulated.cells_at.90min_.ind.                             : num  561 197 230 778 128 ...
#>  $ CLL.B.cell..patient.M7_unstimulated.cells_at.210min_.ind.                            : num  570 210 235 713 128 ...
#>  $ CLL.B.cell..patient.M7_unstimulated.cells_at.390min_.ind.                            : num  595 210 228 647 126 ...
#>  $ CLL.B.cell..patient.M7_stimulated.cells_at.60min_.ind.                               : num  528 187 218 591 134 ...
#>  $ CLL.B.cell..patient.M7_stimulated.cells_at.90min_.ind.                               : num  527 195 217 614 126 ...
#>  $ CLL.B.cell..patient.M7_stimulated.cells_at.210min_.ind.                              : num  653 216 220 737 130 ...
#>  $ CLL.B.cell..patient.M7_stimulated.cells_at.390min_.ind.                              : num  459 211 206 606 140 ...
#>  $ CLL.B.cell..patient.M8_unstimulated.cells_at.60min_.ind.                             : num  503 202 276 621 134 ...
#>  $ CLL.B.cell..patient.M8_unstimulated.cells_at.90min_.ind.                             : num  595 207 303 803 141 ...
#>  $ CLL.B.cell..patient.M8_unstimulated.cells_at.210min_.ind.                            : num  632 200 265 741 143 ...
#>  $ CLL.B.cell..patient.M8_unstimulated.cells_at.390min_.ind.                            : num  740 201 269 813 141 ...
#>  $ CLL.B.cell..patient.M8_stimulated.cells_at.60min_.ind.                               : num  493 192 301 677 132 ...
#>  $ CLL.B.cell..patient.M8_stimulated.cells_at.90min_.ind.                               : num  671 209 269 708 137 ...
#>  $ CLL.B.cell..patient.M8_stimulated.cells_at.210min_.ind.                              : num  721 213 259 848 143 ...
#>  $ CLL.B.cell..patient.M8_stimulated.cells_at.390min_.ind.                              : num  658 206 293 725 134 ...
#>  $ CLL.B.cell..patient.M2_unstimulated.cells_at.60min_.agg.                             : num  701 216 733 790 143 ...
#>   [list output truncated]
# }