Analysing the network

Calculates some indicators for each node in the network.

analyze_network(Omega,nv,...)

Arguments

Omega
a network object

nv
the level of cutoff at which the analysis should be done

...
Optional arguments

Value

A matrix containing, for each node, its betweenness,its degree, its output, its closeness.

Examples

data(network)
analyze_network(network,nv=0)

#> Le chargement a nécessité le package : tnet

#> Le chargement a nécessité le package : igraph

#> 
#> Attachement du package : ‘igraph’

#> The following object is masked from ‘package:Patterns’:
#> 
#>     compare

#> The following object is masked from ‘package:DynDoc’:
#> 
#>     path

#> The following objects are masked from ‘package:stats’:
#> 
#>     decompose, spectrum

#> The following object is masked from ‘package:base’:
#> 
#>     union

#> Le chargement a nécessité le package : survival

#> tnet: Analysis of Weighted, Two-mode, and Longitudinal networks.
#> Type ?tnet for help.

#>     node betweenness degree      output  closeness
#> 1      1         198     45 0.469801103 115.903327
#> 2      2          25     72 0.638928759  83.063082
#> 3      3           4     45 0.650350027  87.061645
#> 4      4         246     56 1.568438860 163.897602
#> 5      5         296     29 0.800405535 147.300670
#> 6      6          22     50 0.631385634  84.830824
#> 7      7         105     40 0.510093629  79.058898
#> 8      8           0     60 0.831951787 101.310557
#> 9      9         773     71 2.467857820 167.135446
#> 10    10         888     73 3.752542041 237.684494
#> 11    11          21     57 1.216543820 181.215213
#> 12    12         112     56 1.204054063 125.028827
#> 13    13         407     58 1.036650559 152.029315
#> 14    14           3     62 0.692952694  85.174289
#> 15    15        1039     56 3.065983757 304.121187
#> 16    16         742     74 3.235406870 236.643385
#> 17    17         418     50 2.008931860 164.416975
#> 18    18         809     43 2.206203179 179.319850
#> 19    19           0     51 0.356455321  59.224482
#> 20    20           0     63 0.549960963  75.273850
#> 21    21           0      1 0.004475704  11.784278
#> 22    22          20     12 0.326518633  98.918177
#> 23    23           0      4 0.006311850  13.858663
#> 24    24           0      0 0.000000000   0.000000
#> 25    25           0      2 0.003019120   7.967493
#> 26    26         186     17 0.095327608  58.848089
#> 27    27           0     33 0.103882029  67.490555
#> 28    28           0      0 0.000000000   0.000000
#> 29    29           0      4 0.056228783  44.016148
#> 30    30           5     10 0.080530134  52.944496
#> 31    31           0      1 0.001812159   6.022486
#> 32    32           0     14 0.074665588  44.828983
#> 33    33           5      8 0.062549930  62.001987
#> 34    34           0      8 0.059090818  56.083944
#> 35    35           1     12 0.142198183  57.742597
#> 36    36           0      0 0.000000000   0.000000
#> 37    37           8     22 0.147872753  57.477842
#> 38    38          21     14 0.223497302  79.456869
#> 39    39          34     26 0.235247646  79.539121
#> 40    40         163     17 0.218449765  95.667232
#> 41    41          11     31 0.211050541  47.312104
#> 42    42           0     22 0.147475862  54.350118
#> 43    43         157     25 0.444488091 139.014044
#> 44    44         176     54 1.026959490 113.149468
#> 45    45           0      8 0.067746335  19.094448
#> 46    46          88     31 1.090647718  94.468427
#> 47    47          52     28 0.636401761  90.551082
#> 48    48         213     35 1.215864042 157.955686
#> 49    49        2253     28 2.158005039 381.339329
#> 50    50         207     30 0.593429905 124.064291
#> 51    51          95     17 0.270615619  95.422431
#> 52    52          28     14 0.188128159  76.876138
#> 53    53         193     38 0.582344605 159.650711
#> 54    54         235     39 1.300831276 198.432375
#> 55    55           0     14 0.178093932  56.611717
#> 56    56           0      2 0.003466005   9.857721
#> 57    57         209     44 1.137639574 158.013782
#> 58    58          10     10 0.252058512  71.745032
#> 59    59           0     11 0.099769365  71.996771
#> 60    60         131     26 0.570673015  87.123197
#> 61    61         157     13 0.389970191 145.794189
#> 62    62           8     13 0.115549109  83.426632
#> 63    63         237     22 0.602373931 180.845454
#> 64    64        1148     56 1.772280889 216.238265
#> 65    65         137     18 0.332703885 123.852149
#> 66    66         100     15 0.105998249  50.620088
#> 67    67        1543     79 6.239688253 304.769647
#> 68    68          56     24 0.659225438  94.288606
#> 69    69        1158     72 7.977272772 392.362109
#> 70    70         666     70 4.786608465 261.311430
#> 71    71         386     36 1.107687884 140.438332
#> 72    72           0     22 0.206933486  84.004642
#> 73    73           0     14 0.154691176  63.244716
#> 74    74         211     15 0.374471030 135.022274
#> 75    75           0      9 0.048956963  57.009683
#> 76    76         478     47 1.518954643 155.919879
#> 77    77          16     30 0.190685214  85.305500
#> 78    78           9     32 0.496323653  79.570820
#> 79    79         226     33 0.777078748 138.700507
#> 80    80          26      6 0.159098923  93.626341
#> 81    81         294     80 4.015011946 203.802868
#> 82    82         323     25 0.653477617 155.237365
#> 83    83         133     14 0.411038467 151.457804
#> 84    84           2     43 0.497062676 107.920114
#> 85    85           0      8 0.102135477  67.458958
#> 86    86         111     37 0.669546198 109.378041
#> 87    87           4     12 0.106633622  64.731765
#> 88    88         234     61 1.366375255 135.747187
#> 89    89          27     26 0.496320464 131.725455
#> 90    90          83     34 0.945657952 106.538998
#> 91    91           0      5 0.014670192  12.148342
#> 92    92         191     22 0.219192961  93.349440
#> 93    93         208     24 0.278074176 158.180808
#> 94    94         163     32 0.736480926 147.682661
#> 95    95          13     10 0.206570745 115.393984
#> 96    96        1204     80 7.977096594 378.317855
#> 97    97           1     20 0.221063915 100.834754
#> 98    98         185     52 0.848957484 102.243228
#> 99    99           0     32 0.168140838  65.515421
#> 100  100         486     63 2.619433306 157.529780
#> 101  101         114     31 0.703401123 236.998887
#> 102  102         434     58 1.549953727 173.419637