See the evolution of the network with change of cutoff

See the evolution of the network with change of cutoff

Usage

# S4 method for class 'omics_network'
evolution(
  net,
  list_nv,
  gr = NULL,
  color.vertex = NULL,
  color.edge = NULL,
  fix = TRUE,
  size = c(2000, 1000),
  label_v = 1:dim(net@omics_network)[1],
  legend.position = "topleft",
  frame.color = "black",
  label.hub = FALSE,
  outdir,
  type.ani = "html"
)

Arguments

net

a omics_network object

list_nv

a vector of cutoff at which the network should be shown

gr

a vector giving the group of each genee. Defaults to NULL

color.vertex

a vector giving the color of each nodee. Defaults to NULL

color.edge

a vector giving the color of each edge. Defaults to NULL

fix

logical, should the position of the node in the network be calculated once at the beginning ? Defaults to TRUE.

size

vector giving the size of the plot. Defaults to c(2000,1000)

label_v

vector giving the labels of each vertex. Defaults to 1:dim(net@omics_network)[1]

legend.position

string giving the position of the legend. Defaults to "topleft"

frame.color

string giving the color of the frame of the plot. Defaults to "black"

label.hub

label hubs. Defaults to FALSE

outdir

Directory to save the animation. No default value since it must be specified by the user.

type.ani

Type of animation. Defaults to "html"

legend

string giving the position of the legend. Defaults to "topleft"

Value

A HTML page with the evolution of the network.

Details

Several types of outputs are available using the type.ani option.

html

latex (requires latex)

swf (requires swftools)

video (requires ffmpeg)

gif

manual_gif

Author

Bertrand Frederic, Myriam Maumy-Bertrand.

Examples

# \donttest{
data(network)
sequence<-seq(0,0.2,length.out=20)

#Change the destdir to have the animation created where you want. 
destdir = tempdir()

#Example of use of the evolution method with an html output.
evolution(network,sequence,type.ani = "html", outdir=destdir)
#> HTML file created at: index.html
# }

if (FALSE) { # \dontrun{
#Example of use of the evolution method with an animated gif output.
evolution(network,sequence,type.ani = "gif", outdir=destdir)
} # }