jetsetscoresHuman transcription factors from HumanTFDB

This data set provides gene target and quality scores for each probe set on the hgu133plus2 Affymetrix gene expression microarrays.

Details

Scores are calculated from BLASTN alignments between probe sequences and Refseq transcript sequences, as described in the vignette and in the reference below.

The Refseq human RNA was downloaded from NCBI on 2017-04-04. The lookups were based on org.Hs.eg.db version 3.4.0.

References

Qiyuan Li, Nicolai J. Birkbak, Balazs Gyorffy, Zoltan Szallasi and Aron C. Eklund. (2011) Jetset: selecting the optimal microarray probe set to represent a gene. BMC Bioinformatics. 12:474.

Author

Bertrand Frederic

Examples

data(jetsetscores)
head(jetsetscores)
#>             nProbes EntrezID process specificity coverage
#> 33323_r_at       16     2810    56.5           1        1
#> 227850_x_at      11   148170    57.0           1        1
#> 33322_i_at       14     2810    70.5           1        1
#> 1566403_at       11    26780    75.0           1        1
#> 201553_s_at      11     3916    80.5           1        1
#> 31637_s_at       16     9572    81.5           1        1
summary(jetsetscores)
#>     nProbes        EntrezID            process         specificity    
#>  Min.   : 8.00   Length:54675       Min.   :   23.0   Min.   :0.0500  
#>  1st Qu.:11.00   Class :character   1st Qu.:  314.5   1st Qu.:0.8182  
#>  Median :11.00   Mode  :character   Median :  466.0   Median :0.9091  
#>  Mean   :11.05                      Mean   : 1315.8   Mean   :0.8524  
#>  3rd Qu.:11.00                      3rd Qu.: 1653.0   3rd Qu.:1.0000  
#>  Max.   :69.00                      Max.   :87998.0   Max.   :1.0000  
#>                                     NA's   :17966     NA's   :17966   
#>     coverage     
#>  Min.   :0.0000  
#>  1st Qu.:0.6667  
#>  Median :1.0000  
#>  Mean   :0.7963  
#>  3rd Qu.:1.0000  
#>  Max.   :1.0000  
#>  NA's   :17966