Plot

Considering the class of the argument which is passed to plot, the graphical output differs.

Methods

signature(x = "micro_array", y = "ANY",...)

x

a micro\_array object

list_nv

a vector of cutoff at which the network should be shown

signature(x = "network", y = "ANY",...)

x

a network object

…

Optionnal arguments:

gr

a vector giving the group of each gene

choice

what graphic should be plotted: either "F" (for a representation of the matrices F) or "network".

nv

the level of cutoff. Defaut to 0.

ini

using the ``position'' function, you can fix the position of the nodes

color.vertex

a vector defining the color of the vertex

ani

vector giving the size of the plot. Default to c(2000,1000). The animation can only be created in the working directory. See the help page of the animation method.

video

if ani is TRUE and video is TRUE, the animation result is a GIF video

label_v

vector defining the vertex labels

legend.position

position of the legend

frame.color

color of the frames

label.hub

logical ; if TRUE only the hubs are labeled

edge.arrow.size

size of the arrows ; default to 0.7

edge.thickness

edge thickness ; default to 1.

signature(x = "micropredict", y = "ANY",...)

x

a micropredict object

…

Optionnal arguments: see plot for network

Examples

if(require(CascadeData)){
data(micro_US, package="CascadeData")
micro_US<-as.micro_array(micro_US[1:100,],time=c(60,90,210,390),subject=6)
plot(micro_US)
}

#> Error in as.double(y): cannot coerce type 'S4' to vector of type 'double'