Returns the position of edges in the network

Returns the position of edges in the network Retrieve network position for consistent plotting. Utility function to plot networks.

Usage

# S4 method for class 'omics_network'
position(net, nv = 0)

Arguments

net

a omics_network object

nv

the level of cutoff at which the analysis should be done

Value

Matrix with as many rows as the number of edges of the network and three columns (name, xcoord, ycoord).

Methods

list("signature(net = \"omics_network\")")

Returns a matrix with the position of the node. This matrix can then be used as an argument in the plot function.

Author

Bertrand Frederic, Myriam Maumy-Bertrand.

Examples

data(network)
position(network)
#>        nom                  
#>   [1,]   1 71.72124 161.7952
#>   [2,]   2 71.63588 160.8057
#>   [3,]   3 71.93108 161.7036
#>   [4,]   4 72.04628 160.9306
#>   [5,]   5 72.11812 160.6961
#>   [6,]   6 72.03740 161.1650
#>   [7,]   7 72.29338 161.3056
#>   [8,]   8 71.81777 161.5850
#>   [9,]   9 71.34111 160.9783
#>  [10,]  10 71.50685 161.1718
#>  [11,]  11 72.10284 161.3768
#>  [12,]  12 71.90836 160.8888
#>  [13,]  13 72.18240 161.1740
#>  [14,]  14 71.90307 161.1251
#>  [15,]  15 71.99710 161.3222
#>  [16,]  16 71.72993 160.9333
#>  [17,]  17 71.57302 161.7361
#>  [18,]  18 72.16648 161.7296
#>  [19,]  19 72.22849 160.9833
#>  [20,]  20 71.91185 161.4455
#>  [21,]  21 71.78632 159.8528
#>  [22,]  22 71.07564 160.6048
#>  [23,]  23 70.25376 161.6431
#>  [24,]  24 70.44650 161.9659
#>  [25,]  25 70.63737 159.8051
#>  [26,]  26 72.17047 161.9519
#>  [27,]  27 72.23753 160.5449
#>  [28,]  28 69.80320 160.7155
#>  [29,]  29 71.42431 160.0008
#>  [30,]  30 70.37108 161.3944
#>  [31,]  31 70.04205 160.3079
#>  [32,]  32 71.04614 162.1410
#>  [33,]  33 70.84640 161.9809
#>  [34,]  34 70.69741 162.7665
#>  [35,]  35 71.75295 160.3139
#>  [36,]  36 69.82547 161.3024
#>  [37,]  37 71.25582 161.7781
#>  [38,]  38 71.24587 162.1620
#>  [39,]  39 70.98649 160.4110
#>  [40,]  40 71.45552 160.5476
#>  [41,]  41 71.67057 162.2483
#>  [42,]  42 72.02734 162.1679
#>  [43,]  43 72.67226 161.3187
#>  [44,]  44 71.03119 161.2100
#>  [45,]  45 72.12445 159.8770
#>  [46,]  46 71.26015 161.9455
#>  [47,]  47 71.47206 162.3105
#>  [48,]  48 72.47406 161.4323
#>  [49,]  49 71.66474 161.9637
#>  [50,]  50 71.15796 161.3350
#>  [51,]  51 71.93856 160.5411
#>  [52,]  52 70.76599 161.7507
#>  [53,]  53 71.79590 160.4912
#>  [54,]  54 72.51372 161.0685
#>  [55,]  55 72.55007 160.2616
#>  [56,]  56 72.09294 162.4191
#>  [57,]  57 71.16171 160.8625
#>  [58,]  58 70.62194 160.6358
#>  [59,]  59 71.77365 162.8129
#>  [60,]  60 71.12385 160.3824
#>  [61,]  61 72.87660 161.0528
#>  [62,]  62 70.63883 160.9937
#>  [63,]  63 72.34776 162.0373
#>  [64,]  64 71.42125 161.6611
#>  [65,]  65 71.45862 162.1782
#>  [66,]  66 72.77947 161.1675
#>  [67,]  67 71.44524 161.4851
#>  [68,]  68 70.77688 161.4036
#>  [69,]  69 71.56672 161.2868
#>  [70,]  70 71.45370 161.3586
#>  [71,]  71 70.93611 161.0694
#>  [72,]  72 72.35003 161.5864
#>  [73,]  73 72.33272 160.1932
#>  [74,]  74 72.62434 161.7640
#>  [75,]  75 72.86204 161.8808
#>  [76,]  76 70.93913 160.8937
#>  [77,]  77 71.91103 161.9671
#>  [78,]  78 71.49422 160.3618
#>  [79,]  79 72.62779 160.9315
#>  [80,]  80 72.69942 160.5317
#>  [81,]  81 71.67846 161.4052
#>  [82,]  82 72.30815 161.7514
#>  [83,]  83 72.50467 162.1998
#>  [84,]  84 70.97550 161.3486
#>  [85,]  85 72.09761 160.2364
#>  [86,]  86 70.96754 161.7201
#>  [87,]  87 72.96419 160.6776
#>  [88,]  88 71.67809 161.0732
#>  [89,]  89 72.46179 160.7181
#>  [90,]  90 71.21330 160.6931
#>  [91,]  91 72.89642 161.5408
#>  [92,]  92 71.83907 162.2871
#>  [93,]  93 71.34313 160.3880
#>  [94,]  94 70.92352 161.4936
#>  [95,]  95 72.27404 162.5184
#>  [96,]  96 71.72321 161.2531
#>  [97,]  97 71.08577 161.7168
#>  [98,]  98 71.26091 161.2191
#>  [99,]  99 72.69758 162.0476
#> [100,] 100 71.45539 160.9974
#> [101,] 101 70.81260 160.7166
#> [102,] 102 71.18860 161.0613