Returns the position of edges in the network

Returns the position of edges in the network Retrieve network position for consistent plotting. Utility function to plot networks.

# S4 method for omics_network
position(net, nv = 0)

Arguments

net

a omics_network object

nv

the level of cutoff at which the analysis should be done

Value

Matrix with as many rows as the number of edges of the network and three columns (name, xcoord, ycoord).

Methods

list("signature(net = \"omics_network\")")

Returns a matrix with the position of the node. This matrix can then be used as an argument in the plot function.

Author

Bertrand Frederic, Myriam Maumy-Bertrand.

Examples

data(network)
position(network)
#>        nom                  
#>   [1,]   1 84.99881 147.9288
#>   [2,]   2 85.68879 147.9407
#>   [3,]   3 85.49458 147.4323
#>   [4,]   4 85.60938 147.5992
#>   [5,]   5 85.75451 147.6253
#>   [6,]   6 85.11131 148.0941
#>   [7,]   7 84.94027 148.1702
#>   [8,]   8 85.28869 147.7782
#>   [9,]   9 85.90622 147.9828
#>  [10,]  10 85.65276 148.2788
#>  [11,]  11 85.10664 147.7029
#>  [12,]  12 85.44447 147.6926
#>  [13,]  13 85.09105 147.8111
#>  [14,]  14 85.36075 147.5554
#>  [15,]  15 85.55320 147.7742
#>  [16,]  16 85.76618 148.1278
#>  [17,]  17 84.80019 148.0478
#>  [18,]  18 85.68809 147.1914
#>  [19,]  19 85.09520 147.4765
#>  [20,]  20 85.22233 147.5841
#>  [21,]  21 86.66128 147.5265
#>  [22,]  22 84.57706 148.7002
#>  [23,]  23 85.81511 149.4428
#>  [24,]  24 86.51116 148.9526
#>  [25,]  25 85.35299 149.7431
#>  [26,]  26 85.45942 147.2595
#>  [27,]  27 85.83448 147.1300
#>  [28,]  28 87.06124 149.1186
#>  [29,]  29 86.45306 148.6291
#>  [30,]  30 86.72143 148.4032
#>  [31,]  31 86.69286 149.4821
#>  [32,]  32 84.80024 148.5917
#>  [33,]  33 86.08049 149.0396
#>  [34,]  34 84.47986 149.5582
#>  [35,]  35 86.17642 148.6900
#>  [36,]  36 86.27945 149.6776
#>  [37,]  37 86.00172 147.6131
#>  [38,]  38 86.27205 147.4628
#>  [39,]  39 85.06473 149.1267
#>  [40,]  40 85.19139 148.9791
#>  [41,]  41 84.49886 148.3314
#>  [42,]  42 84.54461 147.9735
#>  [43,]  43 85.04786 147.1997
#>  [44,]  44 85.69804 148.5647
#>  [45,]  45 84.13011 148.6029
#>  [46,]  46 85.91243 148.7638
#>  [47,]  47 84.66887 148.2442
#>  [48,]  48 84.69404 147.6756
#>  [49,]  49 84.62747 147.8474
#>  [50,]  50 85.69510 148.7464
#>  [51,]  51 84.83778 147.5909
#>  [52,]  52 84.35682 147.2676
#>  [53,]  53 85.83737 147.5184
#>  [54,]  54 85.36967 147.1093
#>  [55,]  55 86.50811 147.3166
#>  [56,]  56 84.64581 146.9170
#>  [57,]  57 86.07195 148.0559
#>  [58,]  58 86.66712 147.9326
#>  [59,]  59 84.15029 148.8376
#>  [60,]  60 85.94461 148.6586
#>  [61,]  61 85.34379 146.8414
#>  [62,]  62 84.40801 147.7700
#>  [63,]  63 85.39834 149.0033
#>  [64,]  64 84.99501 148.3535
#>  [65,]  65 84.39052 148.1532
#>  [66,]  66 86.07273 147.2907
#>  [67,]  67 85.59705 148.0753
#>  [68,]  68 84.97221 148.8885
#>  [69,]  69 85.25867 148.2279
#>  [70,]  70 85.40555 148.3495
#>  [71,]  71 85.88893 148.4385
#>  [72,]  72 85.27740 148.7450
#>  [73,]  73 84.24469 147.6560
#>  [74,]  74 84.38538 147.4658
#>  [75,]  75 86.22912 147.0904
#>  [76,]  76 86.07316 148.2915
#>  [77,]  77 84.71869 147.5001
#>  [78,]  78 85.64697 148.9154
#>  [79,]  79 84.87867 147.0861
#>  [80,]  80 86.54410 147.7136
#>  [81,]  81 85.37875 148.0802
#>  [82,]  82 86.21175 147.9125
#>  [83,]  83 84.88994 146.7274
#>  [84,]  84 85.47460 148.6596
#>  [85,]  85 85.78553 146.7281
#>  [86,]  86 85.08384 148.6625
#>  [87,]  87 84.09620 148.1488
#>  [88,]  88 85.29667 147.9531
#>  [89,]  89 85.16739 146.9912
#>  [90,]  90 86.20097 148.4106
#>  [91,]  91 84.62435 147.1389
#>  [92,]  92 86.18685 147.7062
#>  [93,]  93 84.79677 148.7903
#>  [94,]  94 86.28142 148.1714
#>  [95,]  95 86.02786 146.8868
#>  [96,]  96 85.45924 147.9813
#>  [97,]  97 84.60001 148.5206
#>  [98,]  98 85.49555 148.4957
#>  [99,]  99 85.56458 146.8033
#> [100,] 100 85.52598 148.2665
#> [101,] 101 86.47271 148.1691
#> [102,] 102 85.29527 148.4655