Returns the position of edges in the network

Returns the position of edges in the network Retrieve network position for consistent plotting. Utility function to plot networks.

# S4 method for network
position(net, nv = 0)

Arguments

net
a network object

nv
the level of cutoff at which the analysis should be done

Value

Matrix with as many rows as the number of edges of network and three columns (name, xcoord, ycoord).

Methods

list("signature(net = \"network\")")

Returns a matrix with the position of the node. This matrix can then be used as an argument in the plot function.

Author

Bertrand Frederic, Myriam Maumy-Bertrand.

Examples

data(network)
position(network)

#>        nom                  
#>   [1,]   1 83.05281 145.8135
#>   [2,]   2 83.40773 146.0362
#>   [3,]   3 83.56600 146.7397
#>   [4,]   4 83.57525 145.9787
#>   [5,]   5 83.78106 145.7349
#>   [6,]   6 83.74424 146.0442
#>   [7,]   7 83.90528 145.9258
#>   [8,]   8 83.44175 145.8386
#>   [9,]   9 82.98484 146.6520
#>  [10,]  10 83.43693 146.2960
#>  [11,]  11 83.53720 146.4992
#>  [12,]  12 83.86204 146.2606
#>  [13,]  13 83.63495 146.3657
#>  [14,]  14 83.85369 146.5633
#>  [15,]  15 83.65729 146.1952
#>  [16,]  16 83.24482 146.3316
#>  [17,]  17 83.50506 147.0364
#>  [18,]  18 83.63048 147.3255
#>  [19,]  19 83.59131 145.7576
#>  [20,]  20 83.76982 146.4216
#>  [21,]  21 82.01370 146.4308
#>  [22,]  22 83.10298 147.2948
#>  [23,]  23 82.16892 145.4020
#>  [24,]  24 82.02046 145.7594
#>  [25,]  25 83.78124 147.8160
#>  [26,]  26 83.69376 145.4952
#>  [27,]  27 83.30904 147.3107
#>  [28,]  28 81.78238 145.2170
#>  [29,]  29 82.56562 145.3284
#>  [30,]  30 81.95712 146.1625
#>  [31,]  31 81.53871 145.8285
#>  [32,]  32 82.89107 145.2257
#>  [33,]  33 82.50356 147.0298
#>  [34,]  34 82.86114 147.9982
#>  [35,]  35 83.46860 147.2228
#>  [36,]  36 81.69901 146.9103
#>  [37,]  37 83.02338 147.1432
#>  [38,]  38 83.12490 145.2242
#>  [39,]  39 83.57537 145.3654
#>  [40,]  40 82.70477 147.1654
#>  [41,]  41 83.04784 146.9454
#>  [42,]  42 82.86971 145.4809
#>  [43,]  43 84.05086 146.8024
#>  [44,]  44 82.92224 146.1557
#>  [45,]  45 84.38451 147.3619
#>  [46,]  46 83.00500 145.5592
#>  [47,]  47 83.75205 147.1833
#>  [48,]  48 83.75186 146.9984
#>  [49,]  49 83.18363 145.6699
#>  [50,]  50 83.24677 145.8235
#>  [51,]  51 83.29115 145.2708
#>  [52,]  52 84.23144 147.1738
#>  [53,]  53 83.27580 145.5494
#>  [54,]  54 84.05036 146.1811
#>  [55,]  55 83.96910 145.0635
#>  [56,]  56 82.27613 146.3239
#>  [57,]  57 82.97993 145.9618
#>  [58,]  58 82.29688 146.9424
#>  [59,]  59 83.52515 147.8997
#>  [60,]  60 83.09280 146.6036
#>  [61,]  61 82.58411 146.7388
#>  [62,]  62 82.92435 147.3215
#>  [63,]  63 84.26960 145.7687
#>  [64,]  64 83.39239 146.7624
#>  [65,]  65 82.47474 146.2327
#>  [66,]  66 83.57737 145.1302
#>  [67,]  67 83.24171 146.0958
#>  [68,]  68 82.61192 146.3999
#>  [69,]  69 83.25714 146.6472
#>  [70,]  70 83.09658 146.4312
#>  [71,]  71 82.58702 146.0703
#>  [72,]  72 82.60910 145.7314
#>  [73,]  73 84.63117 146.0159
#>  [74,]  74 84.43989 146.1670
#>  [75,]  75 84.56343 146.6896
#>  [76,]  76 82.79090 145.7999
#>  [77,]  77 83.24049 146.9699
#>  [78,]  78 82.52836 146.5842
#>  [79,]  79 83.98308 145.5126
#>  [80,]  80 84.64778 146.3272
#>  [81,]  81 83.33734 146.5074
#>  [82,]  82 83.91436 146.8720
#>  [83,]  83 84.51945 146.9657
#>  [84,]  84 82.73548 146.1514
#>  [85,]  85 84.01658 147.3762
#>  [86,]  86 82.78399 146.8556
#>  [87,]  87 84.74471 146.4835
#>  [88,]  88 83.64639 146.5994
#>  [89,]  89 84.24795 146.6267
#>  [90,]  90 82.70370 145.9148
#>  [91,]  91 84.45526 145.8777
#>  [92,]  92 84.20494 146.3999
#>  [93,]  93 84.13191 145.4165
#>  [94,]  94 84.14226 146.1168
#>  [95,]  95 82.52971 147.3121
#>  [96,]  96 83.33811 146.2268
#>  [97,]  97 84.10712 145.7291
#>  [98,]  98 82.86076 146.6023
#>  [99,]  99 82.08558 146.6422
#> [100,] 100 83.03994 146.2825
#> [101,] 101 82.36258 146.0340
#> [102,] 102 82.88544 146.4356