Returns the position of edges in the network

Returns the position of edges in the network Retrieve network position for consistent plotting. Utility function to plot networks.

# S4 method for omics_network
position(net, nv = 0)

Arguments

net

a omics_network object

nv

the level of cutoff at which the analysis should be done

Value

Matrix with as many rows as the number of edges of the network and three columns (name, xcoord, ycoord).

Methods

list("signature(net = \"omics_network\")")

Returns a matrix with the position of the node. This matrix can then be used as an argument in the plot function.

Author

Bertrand Frederic, Myriam Maumy-Bertrand.

Examples

data(network)
position(network)
#>        nom                  
#>   [1,]   1 62.64414 143.7356
#>   [2,]   2 62.88878 144.1502
#>   [3,]   3 62.92001 143.8968
#>   [4,]   4 63.18872 144.0912
#>   [5,]   5 63.37619 144.0191
#>   [6,]   6 63.07678 143.8398
#>   [7,]   7 62.89343 143.6090
#>   [8,]   8 63.08808 144.0289
#>   [9,]   9 62.60453 144.6554
#>  [10,]  10 62.64785 143.9310
#>  [11,]  11 63.12154 144.3402
#>  [12,]  12 62.90873 143.7497
#>  [13,]  13 63.14420 144.5876
#>  [14,]  14 63.20240 144.4552
#>  [15,]  15 63.23129 144.2428
#>  [16,]  16 62.66362 144.3456
#>  [17,]  17 63.13169 145.0255
#>  [18,]  18 63.37845 145.1035
#>  [19,]  19 63.23522 143.8439
#>  [20,]  20 63.02699 144.6726
#>  [21,]  21 62.71404 145.5366
#>  [22,]  22 63.02718 143.3660
#>  [23,]  23 61.39171 143.7376
#>  [24,]  24 61.31530 144.1579
#>  [25,]  25 61.31030 145.0677
#>  [26,]  26 63.42551 144.4742
#>  [27,]  27 63.05817 145.3859
#>  [28,]  28 60.86022 144.3256
#>  [29,]  29 61.88191 143.4309
#>  [30,]  30 61.49618 144.6090
#>  [31,]  31 60.93832 143.9231
#>  [32,]  32 62.46224 145.2169
#>  [33,]  33 62.48979 143.1870
#>  [34,]  34 61.41131 145.4644
#>  [35,]  35 63.22692 144.8986
#>  [36,]  36 61.15801 143.4899
#>  [37,]  37 62.80846 145.1898
#>  [38,]  38 62.20456 143.3350
#>  [39,]  39 62.70591 143.1573
#>  [40,]  40 62.07594 143.5982
#>  [41,]  41 62.66023 144.8643
#>  [42,]  42 62.81949 143.4173
#>  [43,]  43 63.47209 144.9223
#>  [44,]  44 62.34465 143.9134
#>  [45,]  45 62.97897 145.6585
#>  [46,]  46 62.34770 143.5091
#>  [47,]  47 62.82929 145.0390
#>  [48,]  48 63.53439 144.2253
#>  [49,]  49 63.44759 144.6928
#>  [50,]  50 62.36893 143.6800
#>  [51,]  51 62.31197 143.1947
#>  [52,]  52 63.82168 143.4117
#>  [53,]  53 62.97689 144.9299
#>  [54,]  54 63.69012 144.3485
#>  [55,]  55 63.08827 142.8509
#>  [56,]  56 63.96496 144.1224
#>  [57,]  57 62.31543 144.1510
#>  [58,]  58 61.89755 145.0958
#>  [59,]  59 64.25897 144.8490
#>  [60,]  60 61.86562 144.3047
#>  [61,]  61 63.87799 144.4742
#>  [62,]  62 63.86460 143.9701
#>  [63,]  63 61.98658 144.6989
#>  [64,]  64 62.81972 144.4764
#>  [65,]  65 63.60083 143.9215
#>  [66,]  66 63.60724 143.3945
#>  [67,]  67 62.78294 144.0389
#>  [68,]  68 61.94974 143.8321
#>  [69,]  69 62.68846 144.5021
#>  [70,]  70 62.65649 144.2233
#>  [71,]  71 62.04294 144.2830
#>  [72,]  72 62.64413 143.4673
#>  [73,]  73 64.02803 144.2742
#>  [74,]  74 63.47171 143.3605
#>  [75,]  75 63.92573 143.6370
#>  [76,]  76 62.12416 144.1679
#>  [77,]  77 62.46253 144.9699
#>  [78,]  78 62.19035 144.8350
#>  [79,]  79 63.65150 143.6969
#>  [80,]  80 64.07886 143.7766
#>  [81,]  81 62.88597 144.3112
#>  [82,]  82 63.35947 143.5684
#>  [83,]  83 64.15462 144.5188
#>  [84,]  84 62.22200 144.0126
#>  [85,]  85 63.67057 145.1801
#>  [86,]  86 62.29380 144.7031
#>  [87,]  87 64.02818 144.8824
#>  [88,]  88 62.92958 144.5783
#>  [89,]  89 63.70530 144.6852
#>  [90,]  90 62.10359 144.4779
#>  [91,]  91 63.78770 145.0114
#>  [92,]  92 61.92849 144.0893
#>  [93,]  93 63.03826 143.1845
#>  [94,]  94 61.93505 144.4828
#>  [95,]  95 62.43588 145.5660
#>  [96,]  96 62.65899 144.0944
#>  [97,]  97 63.26060 143.3039
#>  [98,]  98 62.36622 144.2984
#>  [99,]  99 62.10357 145.2062
#> [100,] 100 62.50989 144.4540
#> [101,] 101 62.02300 144.9081
#> [102,] 102 62.35248 144.4785